+++ components/explore-influencers/InfluencerConnects.tsx
@@ -25,7 +25,7 @@ const InfluencerConnects: React.FC<IProps> = () => {
     return (
 
         <View style={styles.row}>
-            <PremiumActionTag
+            {/* <PremiumActionTag
                 label="Influencers remaining"
                 tooltip={"This means how many influencers you can unlock from the explore influencers page. Please upgrade if you have exhausted the limit here.\n\nLimit recharges every month depending on what plan you are on"}
                 icon="star-four-points"
@@ -45,7 +45,7 @@ const InfluencerConnects: React.FC<IProps> = () => {
                 icon="lightning-bolt"
                 variant="purple"
                 count={connectionCreditsLeft}
-            />
+            /> */}
         </View>
     )
 }
